// types/categories.type.ts
import { categories } from "../constants";

export type Category =
  | "eating"
  | "drink"
  | "credit"
  | "groceries"
  | "utility"
  | "health"
  | "transport"
  | "grooming"
  | "apparel"
  | "entertainment";

export type CategoryIcon =
  | "content-cut"
  | "silverware-fork-knife"
  | "basket"
  | "medical-bag"
  | "glass-wine"
  | "bus"
  | "tshirt-crew"
  | "water"
  | "movie-filter-outline"
  | "credit-card-outline";